Trait ra_db::SourceDatabaseExt

pub trait SourceDatabaseExt: SourceDatabase {
    fn file_text(&self, key0: FileId) -> Arc<String>;

    fn set_file_text(&mut self, key0: FileId, value__: Arc<String>);

    fn set_file_text_with_durability(
        &mut self, 
        key0: FileId, 
        value__: Arc<String>, 
        durability__: Durability
    );

    fn file_relative_path(&self, key0: FileId) -> RelativePathBuf;

    fn set_file_relative_path(&mut self, key0: FileId, value__: RelativePathBuf);

    fn set_file_relative_path_with_durability(
        &mut self, 
        key0: FileId, 
        value__: RelativePathBuf, 
        durability__: Durability
    );

    fn file_source_root(&self, key0: FileId) -> SourceRootId;

    fn set_file_source_root(&mut self, key0: FileId, value__: SourceRootId);

    fn set_file_source_root_with_durability(
        &mut self, 
        key0: FileId, 
        value__: SourceRootId, 
        durability__: Durability
    );

    fn source_root(&self, key0: SourceRootId) -> Arc<SourceRoot>;

    fn set_source_root(&mut self, key0: SourceRootId, value__: Arc<SourceRoot>);

    fn set_source_root_with_durability(
        &mut self, 
        key0: SourceRootId, 
        value__: Arc<SourceRoot>, 
        durability__: Durability
    );

    fn source_root_crates(&self, key0: SourceRootId) -> Arc<Vec<CrateId>>;
}

We don't want to give HIR knowledge of source roots, hence we extract these methods into a separate DB.

Required methods

fn file_text(&self, key0: FileId) -> Arc<String>

fn set_file_text(&mut self, key0: FileId, value__: Arc<String>)

Set the value of the file_text input.

See file_text for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_file_text_with_durability(
&mut self,
key0: FileId,
value__: Arc<String>,
durability__: Durability
)

Set the value of the file_text input and promise that its value will never change again.

See file_text for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn file_relative_path(&self, key0: FileId) -> RelativePathBuf

Path to a file, relative to the root of its source root.

fn set_file_relative_path(&mut self, key0: FileId, value__: RelativePathBuf)

Set the value of the file_relative_path input.

See file_relative_path for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_file_relative_path_with_durability(
&mut self,
key0: FileId,
value__: RelativePathBuf,
durability__: Durability
)

Set the value of the file_relative_path input and promise that its value will never change again.

See file_relative_path for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn file_source_root(&self, key0: FileId) -> SourceRootId

Source root of the file.

fn set_file_source_root(&mut self, key0: FileId, value__: SourceRootId)

Set the value of the file_source_root input.

See file_source_root for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_file_source_root_with_durability(
&mut self,
key0: FileId,
value__: SourceRootId,
durability__: Durability
)

Set the value of the file_source_root input and promise that its value will never change again.

See file_source_root for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn source_root(&self, key0: SourceRootId) -> Arc<SourceRoot>

Contents of the source root.

fn set_source_root(&mut self, key0: SourceRootId, value__: Arc<SourceRoot>)

Set the value of the source_root input.

See source_root for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_source_root_with_durability(
&mut self,
key0: SourceRootId,
value__: Arc<SourceRoot>,
durability__: Durability
)

Set the value of the source_root input and promise that its value will never change again.

See source_root for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn source_root_crates(&self, key0: SourceRootId) -> Arc<Vec<CrateId>>

Loading content...